refactor(customers): move fetch all customers to admin customers controller/services
