+++ backend/src/controllers/core/customers_controller.js
@@ -1,14 +1,5 @@
 import * as customerService from "../../services/core/customer_services.js";
-export async function getCustomers(req, res) {
-  try {
-    const customer = await customerService.getCustomersService();
 
-    res.status(200).json(customer);
-  } catch (error) {
-    console.log("Error retrieving customer", error);
-    res.status(500).json({ message: "Internal Server Error" });
-  }
-}
 export async function getCustomer(req, res) {
   try {
     const { customer, orders } = await customerService.getCustomerService(
