+++ backend/src/services/admin/core/admin_customer_services.js
@@ -1,0 +1,5 @@
+import { Customer } from "../../../model/index.js";
+
+export async function getCustomersService() {
+  return await Customer.find().sort({ createdAt: "asc" });
+}
